+++ tests/examples/clean_after/main.sh
@@ -21,12 +21,14 @@ wait
 [[ -f sub/inp.txt ]] || exit 1
 [[ -f sub/out.txt ]] || exit 1
 
+# Now clean the outputs in sub/, adding an unknown file to test that it is not removed.
 touch sub/unknown
 mkdir other
-(cd other; STEPUP_ROOT="../" stepup clean ../sub/ > ../current_cleanup.txt)
+(cd other; STEPUP_ROOT="../" stepup clean ../sub/ --all --commit > ../current_cleanup.txt)
 
 # Check files that are expected to be present and/or missing.
 [[ -f plan.py ]] || exit 1
 [[ -d sub/ ]] || exit 1
+[[ -f sub/unknown ]] || exit 1
 [[ ! -f sub/inp.txt ]] || exit 1
 [[ ! -f sub/out.txt ]] || exit 1