+++ tests/examples/clean_awaited2/main.sh
@@ -37,7 +37,7 @@ wait
 grep modified output.txt
 
 # Call stepup clean to remove all outputs
-stepup clean . > current_cleanup.txt
+stepup clean --all --commit > current_cleanup.txt
 
 # Check files that are expected to be present and/or missing.
 [[ -f plan.py ]] || exit 1