+++ tests/examples/clean_changed/main.sh
@@ -22,9 +22,16 @@ wait
 # Mofify hello.txt externally to StepUp
 echo changed > hello.txt
 
-# Call stepup clean to remove all outputs
-stepup clean . > current_cleanup.txt
+# Call stepup clean to remove all outputs, which should skip hello.txt
+stepup clean --all --commit > current_cleanup_1.txt
 
 # Check files that are expected to be present and/or missing.
 [[ -f plan.py ]] || exit 1
 [[ -f hello.txt ]] || exit 1
+
+# Call stepup clean to remove all outputs, which should skip hello.txt
+stepup clean --all --commit --unsafe > current_cleanup_2.txt
+
+# Check files that are expected to be present and/or missing.
+[[ -f plan.py ]] || exit 1
+[[ ! -f hello.txt ]] || exit 1