+++ tests/examples/optional_cleanup/main.sh
@@ -22,7 +22,7 @@ wait
 [[ -f test2.txt ]] || exit 1
 
 # Call stepup clean to remove all outputs
-stepup clean . > current_cleanup.txt
+stepup clean --all --commit > current_cleanup.txt
 
 # Check files that are expected to be present and/or missing.
 [[ -f plan.py ]] || exit 1