+++ tests/examples/pending_cleanup/main.sh
@@ -25,7 +25,7 @@ grep elsewhere hello.txt
 grep soon bye.txt
 
 # Call stepup clean to remove all outputs
-stepup clean . > current_cleanup.txt
+stepup clean --all --commit > current_cleanup.txt
 
 # Check files that are expected to be present and/or missing.
 [[ -f plan.py ]] || exit 1